+++ Nix/tmux.nix
@@ -3,6 +3,7 @@
 ## * Use C-a instead of C-b as prefix                                         ##
 ## * Use more natural keybinding for window splitting (- and |)               ##
 ## * Start indexes at 1                                                       ##
+## * Use tmuxinator                                                           ##
 ##                                                                            ##
 ## For an exhaustive list of options, please see `../misc/tmux.conf`          ##
 ##                                                                            ##
@@ -18,6 +19,8 @@ let
 in
 
 {
+  environment.systemPackages = [ pkgs.tmuxinator ];
+
   programs.tmux = if stdenv.isDarwin then {
     enable = true;
     tmuxConfig = mkDefault (tmuxConfig + ''